pyproject.toml
[project]
name = "events2rag"
version = "0.1.0"
description = "Ingest events JSON/ICS feeds and upsert vectors to Qdrant."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dateutil>=2.9.0.post0",
  "requests>=2.32.0",
  "qdrant-client>=1.15.0",
  "sentence-transformers>=3.0.0",
  "icalendar>=6.1.0",
  "recurring-ical-events>=3.5.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.11.0"
]

[project.scripts]
events2rag = "events2rag.main:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

